backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pillow==12.1.0
python-multipart==0.0.22
pydantic==2.12.5
